- Summary: Function SUM returns wrong value for ACNUMERIC data type in test MSSQL based report
- Description of Issue: 
For very large number the result is getting rounded.  

- Description of Resolution: 
Rolling back change so that QA can continue testing.

- TED(s) Resolved:
74196

- Regression ( Yes/No ):
NO

- Code Owner Team: 
Data Access Team

- Code Reviewers: 


- Project ID:
Augusta R2.

- Manual Test Description: 


- Test Automated Cases Executed(Yes/No, if “No”, then explain why):


- Special Notes: 
None

+++ data/org.eclipse.birt.data.aggregation/src/org/eclipse/birt/data/aggregation/impl/TotalSum.java
@@ -123,7 +123,7 @@ public class TotalSum extends AggrFunction
 		 */
 		public Object getSummaryValue( )
 		{
-			return Double.valueOf(sum.doubleValue());
+			return sum;
 		}
 
 	}
